Refactor move files dialog's conversion function to utils
